refactor: split main.py into schemas, bootstrap, routes
main.py was 195 lines mixing Pydantic schemas, MLflow loading, chain
bootstrapping, source serialization, app factory, and routes. Now:

- app/schemas.py: pydantic models (Recommendation*, ActiveModelConfig)
- app/bootstrap.py: parse_active_model_config + load_registered_rag_chain
- app/routes.py: endpoints via APIRouter + serialize_sources
- app/main.py: create_app, lifespan, module-level `app` (45 lines)

test_predict_endpoint.py updated to import ActiveModelConfig from
app.schemas. The mock path "app.main.load_registered_rag_chain" still
resolves since main.py imports the symbol into its namespace.

Addresses smell #5. All 40 unit tests pass.
